// src/lib/api/models/updateSeoTemplateDto.ts
/**
 * Generated by orval v7.17.0 🍺
 * Do not edit manually.
 * RealHub API
 * Multi-tenant Real Estate Platform API
 * OpenAPI spec version: v1
 */
import type { UpdateSeoTemplateDtoPageType } from './updateSeoTemplateDtoPageType';
import type { UpdateSeoTemplateDtoCanonicalRuleJson } from './updateSeoTemplateDtoCanonicalRuleJson';

export interface UpdateSeoTemplateDto {
  pageType?: UpdateSeoTemplateDtoPageType;
  name?: string;
  titleTemplate?: string;
  descriptionTemplate?: string;
  ogTitleTemplate?: string;
  ogDescriptionTemplate?: string;
  canonicalRuleJson?: UpdateSeoTemplateDtoCanonicalRuleJson;
  robotsRule?: string;
}